Reshuffling, nothing to see here.
