+++ Programming Project/bin/.gitignore
@@ -1,2 +1,1 @@
-/.DS_Store
 /uk/
